Started integration of postgresql.

+++ handson/jwt-api/src/model/user.rs
@@ -2,7 +2,7 @@ use serde::Serialize;
 
 #[derive(Debug, Serialize, Clone)]
 pub struct User {
-    pub id: usize,
+    pub id: i32,
     pub username: String,
     pub password: String,
     pub role: String,
